Entity Mapping

+++ src/main/java/com/example/demo/dao/AddressRepo.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.demo.model.Address;
 
-public interface AddressRepo extends JpaRepository<Address, Integer>{
+public interface AddressRepo extends JpaRepository<Address, Long>{
 
 }
