+++ src/main/java/com/example/demo/dao/CommentsRepo.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.demo.model.Comments;
 
-public interface CommentsRepo extends JpaRepository<Comments, Integer>{
+public interface CommentsRepo extends JpaRepository<Comments, Long>{
 
 }
