+++ src/main/java/com/example/demo/dao/ContactRepo.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.demo.model.Contact;
 
-public interface ContactRepo extends JpaRepository<Contact, Integer>{
+public interface ContactRepo extends JpaRepository<Contact, Long>{
 
 }
