+++ src/main/java/com/example/demo/dao/MenuRepo.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.demo.model.Menu;
 
-public interface MenuRepo extends JpaRepository<Menu, Integer>{
+public interface MenuRepo extends JpaRepository<Menu, Long>{
 
 }
