+++ src/main/java/com/example/demo/dao/RatingRepo.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.demo.model.Rating;
 
-public interface RatingRepo extends JpaRepository<Rating, Integer>{
+public interface RatingRepo extends JpaRepository<Rating, Long>{
 
 }
