+++ src/main/java/com/example/demo/dao/RestaurantRepo.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.demo.model.Restaurant;
 
-public interface RestaurantRepo extends JpaRepository<Restaurant, Integer>{
+public interface RestaurantRepo extends JpaRepository<Restaurant, Long>{
 
 }
